requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
email-validator
pyjwt
bcrypt
